aws-glue-home-assignment/requirements.txt
pyathena
PyYAML
smart-open
boto3

